refactor:moved item into seperate component
