+++ src/app/courses/create/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import CoursesBuilder from '@/features/courses/course-builder'
+import CoursesBuilder from '@/features/courses/course_builder/course-builder'
 
 export default function CoursesPage() {
   return <CoursesBuilder />
