+++ src/features/courses/course_builder/course-lesson-input.tsx
@@ -1,0 +1,115 @@
+'use client'
+import RoundedButton from '@/components/ui/button/rounded-button'
+import { FiChevronDown } from 'react-icons/fi'
+import { useState } from 'react'
+
+export default function LessonBlock() {
+  const [lessons, setLessons] = useState<Lesson[]>([
+    { title: '', showContent: false, sections: [] },
+  ])
+  const [showSectionContent, setShowSectionContent] = useState<Record<string, boolean>>({})
+
+  const handleLessonTitleChange = (index: number, value: string) => {
+    const updated = [...lessons]
+    updated[index].title = value
+    setLessons(updated)
+  }
+  const handleSectionChange = (lessonIndex: number, sectionIndex: number, value: string) => {
+    const updated = [...lessons]
+    updated[lessonIndex].sections[sectionIndex] = value
+    setLessons(updated)
+  }
+
+  const handleKeyDown = (e: React.KeyboardEvent, index: number) => {
+    const updated = [...lessons]
+
+    if (e.shiftKey && e.key === 'Enter') {
+      e.preventDefault()
+      updated[index].sections.push('')
+      setLessons(updated)
+    } else if (e.key === 'Enter') {
+      e.preventDefault()
+      updated[index].showContent = true
+      updated.splice(index + 1, 0, { title: '', showContent: false, sections: [] })
+      setLessons(updated)
+    }
+  }
+
+  const sectionHandleKeyDown = (
+    e: React.KeyboardEvent,
+    lessonIndex: number,
+    sectionIndex: number,
+  ) => {
+    if (e.key === 'Enter') {
+      e.preventDefault()
+      const key = `${lessonIndex}-${sectionIndex}`
+      setShowSectionContent((prev) => ({
+        ...prev,
+        [key]: true,
+      }))
+    }
+  }
+  return (
+    <>
+      {lessons.map((lesson, lessonIdx) => (
+        <div key={lessonIdx} className="">
+          {/* Lesson Title Input */}
+          <div className="w-full flex justify-between items-center border-b border-gray-700 mb-2">
+            <input
+              type="text"
+              value={lesson.title}
+              onChange={(e) => handleLessonTitleChange(lessonIdx, e.target.value)}
+              onKeyDown={(e) => handleKeyDown(e, lessonIdx)}
+              placeholder="Add a Lesson title..."
+              className="bg-transparent text-[15px] font-['open_sans'] text-gray-100 placeholder-gray-500 outline-none w-full"
+            />
+            {lesson.showContent ? (
+              <div className="flex w-[150px] m-3">
+                <RoundedButton
+                  rightIcon={<FiChevronDown className="ml-2 w-5 h-5" color="#000000" />}
+                >
+                  Add Content
+                </RoundedButton>
+              </div>
+            ) : (
+              <span className="text-[10px] font-['open_sans'] text-gray-500 ml-2 whitespace-nowrap">
+                Shift + Enter to add new section
+              </span>
+            )}
+          </div>
+
+          {/* Section Inputs */}
+          {lesson.sections.map((section, secIdx) => (
+            <div
+              className="w-full flex flex-col items-end justify-between mb-2"
+              key={`${lessonIdx}-${secIdx}`}
+            >
+              <div
+                key={secIdx}
+                className=" w-[95%] flex justify-between items-center border-b border-gray-700 mb-2"
+              >
+                <input
+                  type="text"
+                  value={section}
+                  onChange={(e) => handleSectionChange(lessonIdx, secIdx, e.target.value)}
+                  onKeyDown={(e) => sectionHandleKeyDown(e, lessonIdx, secIdx)}
+                  placeholder="Add a section..."
+                  className="bg-transparent text-[15px] font-['open_sans'] text-gray-100 placeholder-gray-500 outline-none w-full"
+                />
+                {showSectionContent[`${lessonIdx}-${secIdx}`] && (
+                  <div className="flex w-[160px] m-3">
+                    <RoundedButton
+                      rightIcon={<FiChevronDown className="ml-2 w-5 h-5" color="#000000" />}
+                    >
+                      Add Content
+                    </RoundedButton>
+                  </div>
+                )}
+              </div>
+            </div>
+          ))}
+        </div>
+      ))}
+    </>
+  )
+}
